Change render function declaration
Introduce a macro that merges view struct declaration and function
declaration, in order to locate them adjacent to each other.

Simplify declaration by making palette a global (pending refactor of
color themes) so type signatures can be written out. Got rid of RF(..)
wrapper macro.

Goal is to save keystrokes but keep code obvious. The VIEW macro might
be a bit much, but will let it stand for a bit and see.